Rename reducer file
