+++ webapp/src/reducers/poll_metadata.js
